fixes 1343 and 1344. the hubnet applet is working again, and you can see the error message when you cant connect.
i feel really terrible about all of this code. it seems like all of the client code needs to be rewritten. i may get to that at some point.
theres a lot of duplication between ClientApp and ClientApplet and a lot of really confusing code in there.
also, removed ServerTable and put it inside LoginDialog where it belongs.