Fixing issue with choosers in the teacher client.  This is a HACK fix - see new comment inside ClientInterface.scala.
When a teacher client running inside an applet connected to an activity that has a chooser in it, the teacher client crashed because we attempted to invoke a compiler method, but the applet doesn't have access to the compiler.  This only affects teacher clients running inside applets, not teacher clients running inside the regular HubNet client (the regular HubNet client does have access to some compiler services).

This partially fixes:
https://github.com/NetLogo/HubNet-Webapp/issues/2