Revert "update bench dumps"
This reverts commit 71c56f5dca994cef60476e2735d294954c2a0b2c.

The reverted commit had updated the compiler benchmark dumps because the TestCompileBenchmarks test was failing, but we have since discovered (thanks to Seth!) that this was due to differences between the Eclipse Java compiler and the regular javac. The benchmark dumps were OK, and running `sbt clean` before running the tests fixes the problem.