model runs: make sure value merging is done on a copy
For most Seqs, toArray returns a new WrappedArray. For WrappedArray,
toArray returns a reference to itself. When merging from a Frame that
we were holding on to (i.e., now that we cache frames in the review tab)
the oldValues were getting overwritten... (took me a while to find this one!)