remove special "magic open" support from lexer and language
instead, handle the ___ syntax specially in app.CommandLine, in
GUI-only code

it was a bad choice to have ever made this part of the language in the
first place, and a really, really terrible choice to have made it part
of our lexical syntax (!!!!).  a pox on my past self.