don't mess with java.ext.dirs unless launching GUI
why?  well, we mess with it for JOGL's sake, since we don't want to
pick up a JOGL installation other than the one that comes with
NetLogo.  but if we're not running the GUI, we're not using JOGL.
and messing with java.ext.dirs may have unwanted side effects.
in particular, I discovered today that on JDK 8, Nashorn doesn't
work if java.ext.dirs is empty, because nashorn.jar is in
jre/lib/ext/nashorn.jar under $JAVA_HOME.