split parse0 package off from parse
the idea here is that parse0 only depends on api.  eventually we want
the whole front end to only depend on api, so the Tortoise compiler
won't have any engine dependencies.  but we can only accomplish that
gradually, so we'll do that by gradually move things from parse to
parse0.  to start with, it's just the first 2/3 of StructureParser

StructureDeclarations is in its own file now since we plan to start
using it in subsequent phases of compilation.  while I was at it,
I thought it made sense to make the other pieces of StructureParser
stand alone in their own source files, too