remove support for 'permissive' parsing
where permissive means that unknown identifiers didn't causing parsing
to abort with an error, but just be flagged as _unknownidentifier.

this was only used in an obscure and little-used part of the GUI,
namely input boxes where the input type was "command" or "reporter".
it's conceivable it might have proved useful for something else
someday, but that's a bridge we can cross when we come to it, if
we ever come to it.  at the moment, it's just getting in the way
as I refactor.