eliminate TokenType.Variable
there is no need to distinguish built-in agent variables from other
identifiers at the lexing stage. the code for handling it already
existed in IdentifierParser, so we can just wait and let it be handled
then.  yay one fewer token type, yay simplicity!

note that tokenizeForColorization is currently broken, because lots of
things end up as TokenType.Ident instead of being classified as
TokenType.{Command, Reporter}.  but nobody's using it and there no
tests for it; maybe I'll un-break it later