rename TokenType.Constant to TokenType.Literal
it's a peculiarity of NetLogo that named constants like `e` and `pi`
are actually literals rather than 0-argument reporters.  (in the old
days we didn't have ConstantFolder, so there would have been a
performance price to pay for having them be reporters.)  but I think
it's clearer, less surprising and more standard, namingwise, to
subsume named constants as a special case of literals, rather than
vice versa.