remove 'tokenizeForColorization' support
I could definitely see this being added back at some point.  or maybe
not -- Jason already has a JavaScript-based colorizer.  anyway, if
it's ever brought back, it should be done by running Tokenizer plus
whatever the relevant pieces of IdentifierParser are

this is worth getting rid of right now because it's getting the
way of changing how TokenLexer, Tokenizer, and IdentifierParser
interrelate.  in particular, I'm interested in splitting
IdentifierParser up into pieces and having some of them run before
StructureParser and/or ExpressionParser some of them after; that
kind of large-scale change is easier to make with less stuff in
the way