store JFlex-generated sources in standard sbt location
for a long time we've been getting away with having them mixed in
with the rest of our sources, but I'm finding that with sbt 0.13
and Scala 2.11 I get problems with sbt seeing imaginary multiple
copies of the same files and reporting compile errors as a result.

it's probably better to follow the standard sbt structure anyway,
so I'm going ahead and comitting this on the headless branch even
though we don't have either sbt 0.13 or Scala 2.11 there yet.