use '/usr/bin/env scala' to start our auxiliary scala scripts
it's gotten annoying to maintain the 'bin/scala' script in the NetLogo
repo, so let's just stop. the original ideas behind that script were:
1) nobody had to install scala at the system level in order to do
   anything
2) the scripts always got the right version of Scala

now that we're using sbt not make to build, point 1 is already mostly
taken care of. so it no longer seems worth it to me to maintain the
'bin/scala' thing if it's only used by scripts that aren't essential
to the build.

as for point 2, changes in Scala have settled down a lot, so I'm
not especially worried that little-used scripts might break from
time to time. it seems like a bridge easily crossed when we come
to it.

a few other scripts, bin/profiles.scala, and bin/benches.scala,
are more complicated because they need NetLogo on the classpath.
I will handle them in separate commits.