make bin/shell.sh use sbt to launch
I hadn't done it this way before because of JLine issues, but
I found a way to eliminate those.

the old way was hard to maintain because it had version number
and classpath stuff hardcoded into it, and it required `sbt package`
to be manually run first, which was easy to forget to do, and then
it either wouldn't work or you'd be running out of date code

the downside of going through sbt is increased startup time, but
oh well.