give all report/report_* methods explicit return types
the main impetus for this was that in Scala 2.11, if you override
an AnyRef-returning method the compiler will no longer infer a more
specific return type if it can. the more specific return types are
valuable because the bytecode generator can use them.

committing this even for use with 2.10 because it seems like a
code improvement, actually; better not to rely on type inference
for something so crucial as the return types of these methods.

the benchmark dumps change a little, for the better, because some
int-to-double conversions are now happening before bytecode generation