mirroring/Tortoise: don't cache AgentKeys
this is a prospective fix for NetLogo/Tortoise#1

even assuming this change turns out to do the trick (we won't know
until after a string of successful Travis runs), note that it does
substantially increase memory usage for recording model runs. after
all, reducing memory usage is the reason we added the caching in the
first place.  so we may end up confronting this again.