replace coffeescripted-sbt plugin (it was too slow)
it was too slow because it used Rhino.  so instead, use custom code
that installs the right version of the CoffeeScript compiler in the
local node_modules directory and then runs it as external process. way
faster.

as a bonus, we get a newer CoffeeScript version (1.6.3 instead of 1.2.0)

fixes NetLogo/Tortoise #32
fixes NetLogo/Tortoise #33