model runs: extend use of enums for variables to every kind of mirrorable
This required getting rid of the storage of interface globals in the Observer
mirrorable, because the number of those varies from model to model.

They are not going to be needed anyway because we are going to mirror the
widget values as separate mirrorables anyway (that was the original purpose
of having interface globals there, but it would not have worked because some
extra information is needed, e.g. slider constraints)

Cherry pick and conflict resolution of f3cb262