review tab: halt running model when switching run
Previously, if you tried to switch to a run of the same model while recording,
a new run, starting at the point where the model currently is, would be
created (and selected!) right away, effectively preventing you from switching
runs while a model is running.

Switching to a different model didn't have that problem because recording
is stopped anyway when unloading a model.