remove 'private' from some compile.middle classes
I want to use Optimizer from Tortoise.  that's the immediate
motivation, but also this is just a good change.  now that
compile.middle is its own package, if somebody wants to forbid
depending on it, that should be forbidden via Classycle, rather
than forbidden via `private`

if someone just wants to use the middle end as a whole, we already
have `MiddleEnd extends MiddleEndInterface`, so that's taken care
of.