activate URLs in Model Library dialog left pane, fix #751
Note that this is *only* for the general introductory text of the model library. Description of individual models (coming from the WHAT IT IS section of the info) have their HTML explicitely stripped out, so any link in there appears as plain text.

It would be a separate issue if we wanted to fix this.