use absolute path in `AbstractWorkspace.convertToNormal()`
This is used by the `__edit` primitive and wasn't working anymore since the
changes to the directory structure in hexy.

I have also removed the `isDirectory()` check on `.git` because we now look
for the `.git` file in models instead of the NetLogo's root, and that one is
a symlink to `../.git/modules/models` instead of being a proper directory.