Bugfix: Clear Gaussian var in RNG when setting seed
This was causing problems in NetLogo Web, because of the saved RNG
state.  When Headless opened a new model, it got a new `Workspace`
(and, thus, a new RNG).  When NLW opened a new model, it continued
using the same old RNG.

Ideally, it shouldn't matter--new or old--because each model is run
only after doing `random-seed 0`.  However, setting the seed wasn't
clearing the variable that this commit clears, so, when the RNG state
was dumped out to assess whether the RNGs were the same, there
would be a state mismatch for this variable that is functionally
(and currently, at least) irrelevant in its value (when `__have...`
is `false`).

So, while the **behaviors** of the RNGs were correct, the state dump
was... not entirely false, but problematic.  This change makes the state
dumps line up.