Bugfix: Prevent exception on background click in ExtMan in macOS
In macOS clicking on the background spaces of ExtMan would throw an exception due to the
`JTabbedPane` being set to a `CardLayout`.  As we only have 1 tab at the moment, I set
the library dialog to use a standard `JPanel` for now to avoid the error, and we can
switch back to a real `JTabbedPane` once we have 2 or more libraries to display.