Bugfix: Do not `realloc()` world when code tab compile fails
If you `realloc()` the world when the compile fails, some assumptions
the world makes will be incorrect, like there being breeds setup for
links when there are not without a successful compile.

In order to preserve the fix for wiping slider values when reloading a
model with a code tab error (which introduced this regression), I added
a call to the `setGlobalVariables()` on a failed compile.  It seems to
preserve the wiping slider values fix.

Fixes #1758