Bugfix: Give correct topology regions with wrapping for `in-radius`
This should actually fix #1763.  The core problem was when wrapping
around in a way that had overlap, the "higher" y-axis regions would be
generated after the "lower" ones, but the code didn't account for the
"lower" being higher than the "higher" due to said wrapping.

Also made an attempt to add some tests to verify `in-radius`
functionality in general by running a naive NetLogo-code version against
the built-in prim and checking the results.