Infrastructure: Do a git clean before build/test in Jenkins
Jenkins re-uses directories per branch for build and test, it does not
clean them out between commits.  This works fine most of the time, but
when there are tooling updates, like sbt, that leave their own temp
directories around it can foul thigns up.  So it's best to do a full
clean of the repo and start from scratch each time.