Bugfix: `InterfacePanelLite` widgets need to raise added/remove events
The prior fix for #1476 worked for the main NetLogo GUI but didn't take
into account the `InterfacePanelLite` used by HubNet and LevelSpace,
which does not use the `WidgetWrapper` class that the fix relied on.  So
we roll back that fix and implement it differently, allowing the
non-widget widgets to override the event raising code that would cause
the original "dirty" issue.