API: Use Java-compatible default parameters for `open`
This probably should've been fixed a while ago, but better late than
never.  Java does not understand Scala default parameters, so adding a
parameter to an existing method with a default is a breaking change for
any existing Java code that relies on it.  Not a problem for active
projects which can just add the parameter, but for unmaintained projects
it's a needless break when we can just do it the "old fashioned" way.