Fix auto-reload breaking on load errors
If NetLogo encounters an error reloading the model, auto-reload can't be
triggered again until the model is re-opened. This is because Tabs sets
a flag to block auto-reloads if one is already in-progress. Since it
never gets notified when an error occurs, the flag will remain set
indefinitely. Fix this by defining a new event LoadErrorEvent, raise it
when an error occurs, and unset the flag when the event is raised.