Refactor libraries code in StructureParser
In the section of the code where libraries are handled, there's already
a check to ensure that the list of libraries is not empty. Because of
this, it's not necessary to use headOption to access the head of the
list safely. This replaces headOption to head where applicable and apply
some simplifications.