Feature: Change wording and link around Community Models page
We would like to replace this page with something newer, better, and more in-line with our newer visual design language.  What, specifically, that will be, we're not certain.  So I softened the language here, and changed the URL to a page that currently redirects to the same old Community Models page, but will probably redirect to something else in the future.