Refactor/Minor: Use `LinkedHashSet` for jump line trail drawing
In theory a regular ol' `Set` should be fine for the jump line drawer
calcs instead of a `LinkedHashSet`.  In practice, the `Set` gives a
different order for the lines to draw (it's not insert-ordered).  This
shouldn't affect what happens in NetLogo, as the lines for each turtle
are calc'd in order, but when we're trying to verify NetLogo Web works
identically, this becomes an issue.  The simplest solution seems to be
using the slightly different data structure to get the same effect.