Disable Check button and display warning label when included file is not compilable
If an included file has yet to be saved or is not referenced in the main Code tab, it can't be compiled. Prior to this commit, the check button would still be enabled in those situations, but nothing would be compiled and no errors would be generated regardless of the contents of the tab. This fix makes it more clear to the user what's going on.