Bugfix: Ensure color picker alpha slider gets set properly
This was broken by adding the functionality that transferred colors from
the "Simple" tab to the "Advanced" tab.  When we do that, we set the
"Advanced"'s color to that of the "Simple", which doesn't support alpha.
As a result, all alpha information is lost by switching tabs.

That's *maybe* fine, but was particularly problematic, because we were
"switching tabs" *after* setting the initial color value in the
"Advanced" picker.  Which meant that, at the end of launching the color
picker, it wasn't possible for the alpha slider to have any value other
than 100%.