Track whether users accept or decline analytics
Jason says: I requested this change so we can at least know how
representative our analytics and download data actually are, by checking
how many people are truly opening the application after downloading and
making the choice about whether they're okay with tracking or not.  By
making an HTTP request to the server, we theoretically *are* handing the
server identifying information about *everyone*, but the server does not
record that info anywhere that I have seen.  All that it records are the
time of the request and whether or not the requester OKed tracking (i.e.
`enabled`).